Moved make_run_dir to utils
